more progress on converting to using the nextGenConfig.json file
