--- commands/generate/pages/index/indexUtils.js
@@ -1,1 +1,0 @@
-"use strict";
